llm-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
torch
transformers
peft
faiss-cpu
sentence-transformers
rank-bm25
numpy